+++ dubbo-samples-async-generated-future/src/main/java/org/apache/dubbo/samples/async/EmbeddedZooKeeper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.samples.async;
+package org.apache.dubbo.samples.async;
 
 import org.apache.zookeeper.server.ServerConfig;
 import org.apache.zookeeper.server.ZooKeeperServerMain;
