+++ dubbo-samples-async-generated-future/src/main/java/org/apache/dubbo/samples/async/impl/GreetingsServiceImpl.java
@@ -17,17 +17,25 @@
  *
  */
 
-package com.alibaba.dubbo.samples.async;
+package org.apache.dubbo.samples.async.impl;
 
-import org.springframework.context.support.ClassPathXmlApplicationContext;
+import org.apache.dubbo.samples.api.GreetingsService;
 
-public class AsyncProvider {
+/**
+ * AsyncServiceImpl
+ */
+public class GreetingsServiceImpl implements GreetingsService {
 
-    public static void main(String[] args) throws Exception {
-        new EmbeddedZooKeeper(2181, false).start();
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"META-INF/spring/async-provider.xml"});
-        context.start();
-        System.in.read();
+    @Override
+    public String sayHi(String name) {
+        System.out.println("provider received: " + name);
+        try {
+            Thread.sleep(500);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+        System.out.println("provider returned.");
+        return "hello, " + name;
     }
 
 }
